requirement.txt
django-weasyprint
pymupdf
pillow
